course reports arch + csv helper
