+++ supabase/functions/get-course-reports/getCourseReports.ts
@@ -1,18 +1,6 @@
 import EdgeFunctionRequest from "../_shared/EdgeFunctionRequest.ts";
 import { getRows } from "../_shared/database.ts";
-
-/**
- * Converts an array of objects (with the same keys & no embedded objects) into a CSV string
- */
-const toCSV = (json: { [key: string]: any }[]) => {
-    let csv = "";
-    const keys = (json[0] && Object.keys(json[0])) || [];
-    csv += keys.join(',') + '\n';
-    for (let line of json) {
-        csv += keys.map(key => line[key]).join(',') + '\n';
-    }
-    return csv;
-}
+import { toCSV } from "../_shared/helpers.ts";
 
 const getCourseReports = (request: EdgeFunctionRequest) => {
